backend/requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
tzdata==2023.3
djangorestframework==3.14.0
requests==2.31.0
Pillow==10.1.0
gunicorn==20.0.4
django-cors-headers==4.3.0
whitenoise==6.6.0